Starting to revisit this project.
I want to implement a Screen object which is just a list of buttons etc which will be on that screen and then I can display/hide a screen.
This will be mainly used for menus.  Show a menu screen, press a button, hide the menu screen, show the next screen and so on.

Also started to test out the idea that I could draw the screens faster by not display.update every small change, instead pass all the changes as
a list to display.update which is supposed to be faster.